Fix: Formatting, Compile Warnings/Errors (multi-platform)
FossilOrigin-Name: 3226848def55d4284353d4a7c63ebe6dd8d0295e345eab482c32f5fe8896f629

+++ src/jcrypto/sign.c
@@ -46,7 +46,7 @@ Janet cfun_sign(int32_t argc, Janet *argv) {
         crypto_panic_ssl("failed to sign");
     }
 
-    Janet result = janet_stringv(sig, siglen);
+    Janet result = janet_stringv(sig, (int32_t)siglen);
     janet_free(sig);
     EVP_PKEY_free(pkey);
     EVP_MD_CTX_free(mdctx);
